+++ client/src/page/admin/grade/TableData.jsx
@@ -21,16 +21,21 @@ const TableData = ({ setDetail }) => {
     useDeleteGradeMutation();
 
   const deleteHandler = (id) => {
-    toast.promise(
-      deleteGrade(id)
-        .unwrap()
-        .then((res) => res.message),
-      {
-        loading: "Memproses...",
-        success: (message) => message,
-        error: (err) => err.data.message,
-      }
+    const confirm = window.confirm(
+      "Apakah anda yakin ingin menghapus data tingkat dan semua data yang terkait dengan tingkat ini?"
     );
+    if (confirm) {
+      toast.promise(
+        deleteGrade(id)
+          .unwrap()
+          .then((res) => res.message),
+        {
+          loading: "Memproses...",
+          success: (message) => message,
+          error: (err) => err.data.message,
+        }
+      );
+    }
   };
 
   useEffect(() => {
@@ -53,37 +58,37 @@ const TableData = ({ setDetail }) => {
       totalPages={totalPages}
       isLoading={dataLoading}
     >
-      <table className='table table-bordered table-striped table-hover'>
+      <table className="table table-bordered table-striped table-hover">
         <thead>
           <tr>
-            <th className='text-center'>No</th>
-            <th className='text-center'>Satuan</th>
-            <th className='text-center'>Tingkat</th>
-            <th className='text-center'>Aksi</th>
+            <th className="text-center">No</th>
+            <th className="text-center">Satuan</th>
+            <th className="text-center">Tingkat</th>
+            <th className="text-center">Aksi</th>
           </tr>
         </thead>
         <tbody>
           {grades?.map((item, i) => (
             <tr key={i}>
-              <td className='text-center align-middle'>
+              <td className="text-center align-middle">
                 {(page - 1) * limit + i + 1}
               </td>
-              <td className='align-middle'>{item.homebase}</td>
-              <td className='text-center align-middle'>{item.name}</td>
-              <td className='text-center align-middle'>
-                <div className='d-flex justify-content-center gap-2'>
+              <td className="align-middle">{item.homebase}</td>
+              <td className="text-center align-middle">{item.name}</td>
+              <td className="text-center align-middle">
+                <div className="d-flex justify-content-center gap-2">
                   <button
-                    className='btn btn-sm btn-warning'
+                    className="btn btn-sm btn-warning"
                     onClick={() => setDetail(item)}
                   >
-                    <i className='bi bi-pencil-square'></i>
+                    <i className="bi bi-pencil-square"></i>
                   </button>
                   <button
-                    className='btn btn-sm btn-danger'
+                    className="btn btn-sm btn-danger"
                     disabled={isLoading}
                     onClick={() => deleteHandler(item.id)}
                   >
-                    <i className='bi bi-folder-minus'></i>
+                    <i className="bi bi-folder-minus"></i>
                   </button>
                 </div>
               </td>
