+++ client/src/page/admin/teacher/TableData.jsx
@@ -21,16 +21,21 @@ const TableData = ({ setDetail }) => {
     useDeleteTeacherMutation();
 
   const deleteHandler = (id) => {
-    toast.promise(
-      deleteTeacher(id)
-        .unwrap()
-        .then((res) => res.message),
-      {
-        loading: "Memproses...",
-        success: (message) => message,
-        error: (err) => err.data.message,
-      }
+    const confirm = window.confirm(
+      "Apakah anda yakin ingin menghapus data guru ini dan semua data yang terkait dengan guru ini?"
     );
+    if (confirm) {
+      toast.promise(
+        deleteTeacher(id)
+          .unwrap()
+          .then((res) => res.message),
+        {
+          loading: "Memproses...",
+          success: (message) => message,
+          error: (err) => err.data.message,
+        }
+      );
+    }
   };
 
   useEffect(() => {
@@ -53,56 +58,56 @@ const TableData = ({ setDetail }) => {
       totalPages={totalPages}
       isLoading={dataLoading}
     >
-      <table className='m-0 table table-bordered table-striped table-hover'>
+      <table className="m-0 table table-bordered table-striped table-hover">
         <thead>
           <tr>
-            <th className='text-center'>No</th>
-            <th className='text-center'>Satuan</th>
-            <th className='text-center'>Username</th>
-            <th className='text-center'>Nama</th>
-            <th className='text-center'>Jenis Kelamin</th>
-            <th className='text-center'>Mata Pelajaran</th>
-            <th className='text-center'>Wali Kelas</th>
-            <th className='text-center'>Aksi</th>
+            <th className="text-center">No</th>
+            <th className="text-center">Satuan</th>
+            <th className="text-center">Username</th>
+            <th className="text-center">Nama</th>
+            <th className="text-center">Jenis Kelamin</th>
+            <th className="text-center">Mata Pelajaran</th>
+            <th className="text-center">Wali Kelas</th>
+            <th className="text-center">Aksi</th>
           </tr>
         </thead>
         <tbody>
           {teachers?.length > 0 ? (
             teachers?.map((teacher, i) => (
               <tr key={i}>
-                <td className='text-center align-middle'>
+                <td className="text-center align-middle">
                   {(page - 1) * limit + i + 1}
                 </td>
-                <td className='text-center align-middle'>{teacher.homebase}</td>
-                <td className='text-center align-middle'>{teacher.username}</td>
-                <td className='align-middle'>{teacher.name}</td>
-                <td className='text-center align-middle'>
+                <td className="text-center align-middle">{teacher.homebase}</td>
+                <td className="text-center align-middle">{teacher.username}</td>
+                <td className="align-middle">{teacher.name}</td>
+                <td className="text-center align-middle">
                   {teacher.gender == "L" ? "Laki Laki" : "Perempuan"}
                 </td>
-                <td className='align-middle'>
+                <td className="align-middle">
                   {teacher.subjects?.map((item) => (
-                    <p key={item.id} className='m-0'>
+                    <p key={item.id} className="m-0">
                       {item.name}
                     </p>
                   ))}
                 </td>
-                <td className='text-center align-middle'>
+                <td className="text-center align-middle">
                   {teacher.homeroom ? teacher.class_name : "-"}
                 </td>
-                <td className='text-cente align-middle'>
-                  <div className='d-flex justify-content-center gap-2'>
+                <td className="text-cente align-middle">
+                  <div className="d-flex justify-content-center gap-2">
                     <button
-                      className='btn btn-sm btn-warning'
+                      className="btn btn-sm btn-warning"
                       onClick={() => setDetail(teacher)}
                     >
-                      <i className='bi bi-pencil-square'></i>
+                      <i className="bi bi-pencil-square"></i>
                     </button>
                     <button
-                      className='btn btn-sm btn-danger'
+                      className="btn btn-sm btn-danger"
                       disabled={isLoading}
                       onClick={() => deleteHandler(teacher.id)}
                     >
-                      <i className='bi bi-folder-minus'></i>
+                      <i className="bi bi-folder-minus"></i>
                     </button>
                   </div>
                 </td>
@@ -110,7 +115,7 @@ const TableData = ({ setDetail }) => {
             ))
           ) : (
             <tr>
-              <td colSpan={8} className=''>
+              <td colSpan={8} className="">
                 Data Belum tersedia
               </td>
             </tr>
